+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 0.9.0 - 2024-10-03
+
+### Added
+
+- Visibility of Changelog in the game settings
+- Visibility of asset attribution in the game settings
+
+### Changed
+
+- Icon of 'zen' mode is updated, for better contrast
+
 ## Version 0.8.0 - 2024-10-01
 
 ### Added